refactor(kernel,utils): Move counter to ocap utils.
