+++ packages/utils/src/counter.ts
